agl-compositor_git.bb: Bump agl-compositor SRCREV
This SRCREV bump update is necessary to avoid a circular dependency
between libweston and the AGL compositor. This update includes Xwayland
support (and a few fixes related to that) and an update for using
logging scopes to match with that we have in upstream.

Note that Xwayland requires to build with Xwayland enabled and have a
working Xwayland image (which requires a bit of tweaks to make it work).

In more detail, the following changes have been added since previous
bump:

- a68408367cf desktop: Fix Xwayland build
- 5ec94158612 compositor: Use the logging context when iterating over scopes
- f3ae1a25e24 layout: Check against app_id being valid
- 26700fa20ab layout: Determine xwayland surface in a generic fashion
- 8033c46081b compositor: Add XWayland basic support

Change-Id: Ib4aa552eec776a5badafb56bbdd2f780eb0ff331
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/29101
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>